platform-app/requirements.txt
pyqt5==5.15.10
pyqtwebengine==5.15.6
sqlalchemy>=2.0.36
bcrypt==4.1.2
pyjwt==2.8.0
reportlab==4.0.9
jinja2==3.1.3
markdown==3.5.2
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests>=2.31.0
